+++ changelog.md
@@ -1,3 +1,4 @@
+* Offer systemd service file
 * Download and import [Game Backup Monitor](https://github.com/MikeMaximus/gbm) XML file
 * Detect ambiguous game entries
 * Distinguish ambiguous game entries by full path