+++ changelog.md
@@ -1,3 +1,4 @@
+* Allow to run commands when game got started or closed
 * Offer systemd service file
 * Download and import [Game Backup Monitor](https://github.com/MikeMaximus/gbm) XML file
 * Detect ambiguous game entries